+++ cli/src/commands/fix.rs
@@ -359,7 +359,13 @@ async fn fix_one_file(
             &prev_content,
             &extra_args,
         ) {
-            Ok(next_content) => next_content,
+            Ok(next_content) => {
+                if next_content.is_empty() && tool_config.ignore_empty_output {
+                    prev_content
+                } else {
+                    next_content
+                }
+            }
             // TODO: Because the stderr is passed through, this isn't always failing
             // silently, but it should do something better will the exit code, tool
             // name, etc.
@@ -510,8 +516,14 @@ struct ToolConfig {
     /// diffs (e.g., to sort imports, or run with
     /// `--include-unchanged-files`).
     run_tool_if_zero_line_ranges: bool,
-    // TODO: Store the `name` field here and print it with the command's stderr, to clearly
-    // associate any errors/warnings with the tool and its configuration entry.
+    /// Whether to skip applying the tool if its stdout is completely empty. The
+    /// default is `true`, so a tool that prints nothing as a "no-op" won't wipe
+    /// the contents of a file. However, if empty output is desired or expected,
+    /// this can be set to `false`.
+    ignore_empty_output: bool,
+    // TODO: Store the `name` field here and print it with the command's stderr,
+    // to clearly associate any errors/warnings with the tool and its
+    // configuration entry.
 }
 
 /// Represents the `fix.tools` config table.
@@ -533,9 +545,15 @@ struct RawToolConfig {
     line_range_arg: Option<String>,
     #[serde(default)]
     run_tool_if_zero_line_ranges: bool,
+    #[serde(default = "default_ignore_empty_output")]
+    ignore_empty_output: bool,
 }
 
 fn default_tool_enabled() -> bool {
+    true
+}
+
+fn default_ignore_empty_output() -> bool {
     true
 }
 
@@ -574,6 +592,7 @@ fn get_tools_config(
                 enabled: tool.enabled,
                 line_range_arg: tool.line_range_arg,
                 run_tool_if_zero_line_ranges: tool.run_tool_if_zero_line_ranges,
+                ignore_empty_output: tool.ignore_empty_output,
             })
         })
         .try_collect()?;